Support subsets, implement getopt interface
This change also introduces support for selecting the downloaded font
formats, selecting the subset of the fonts to be downloaded and
selecting the output file (including stdout, which can be achieved
using -). It also allows specifying the font names and weights on the
command line instead of by editing the file.