Check getopt(1) version, fix usage note on misuse
Check whether the version of getopt(1) is the GNU one that supports
preserving whitespace and special characters and long options, or the
BSD one that doesn't support either (as is the default on OS X, for
example). Fall back to short options only and print a warning if the old
version is detected.

Additionally, since set -e is used, deal with getopt(1) errors in a way
that actually prints the usage information as originally intended rather
than exiting right away after getopt's message.

Closes: #1