tests: Use default target in test/ to force -k
Even if one of the tests fails, we still want to run all others, too.
The default make target in test/Makefile re-runs make with -k to achieve
that. Let's avoid calling the test target directly to also get the same
treatment when calling make test at the top level.