Sort WOFF2 before WOFF by default, recommend WOFF2
Apparently browsers process the list of fonts in-order and stop at the
first supported option. For this reason, it does not make sense to put
WOFF2 after WOFF when generating both, as reported in #21. Instead,
change the order of the 'all' default format to start with WOFF2 and
WOFF, sort the older formats towards the end and change the
recommendation to use WOFF2 and WOFF.

Closes: #21